+++ util/tracing/detect/detect.go
@@ -79,10 +79,6 @@ func getExporter() (sdktrace.SpanExporter, error) {
 		return nil, err
 	}
 
-	if exp != nil {
-		exp = makeThreadSafe(exp)
-	}
-
 	if Recorder != nil {
 		Recorder.SpanExporter = exp
 		exp = Recorder
